requirements.txt
beautifulsoup4==4.11.1
eyed3==0.9.6
requests==2.28.0
python-dotenv==0.20.0
pyTelegramBotAPI==4.6.0
